Added Command line file
added a .py script that can be called in the command line. it lexes, parses, and executes a ARBITER script.

Added example.txt, which has a ARBITER script that multiplies 2 and 10. it shows how looping takes different amounts of time depending on which input you use as a counter

use => .\ARBITER.py example.txt <= to execute the example program. this will execute it and dump out every variable space on every tick